Revert IntLiteralType to use bigint. Emit RationalLiteralType in infer.ts when encountering rationals. Move evalConstExpr logic to eval_const.ts
