+++ src/types/ast/int_literal.ts
@@ -1,38 +1,20 @@
-import Decimal from "decimal.js";
 import { Range } from "../../misc";
 import { smallestFittingType } from "../utils";
 import { IntType } from "./int_type";
-import { TypeNode } from "./type";
+import { NumericLiteralType } from "./numeric_literal";
 
-export class IntLiteralType extends TypeNode {
+export class IntLiteralType extends NumericLiteralType {
     /// TODO(dimo): After removing the typestring parser make this required
-    /// TODO(dimo): Made a mistake - should revert this to bigint and then
-    /// use the rational_literal type to model non-whole numbers.
-    /// TODO(dimo): Perhaps use a differnt library other than Decimal that keeps explicit track of
-    /// bigint numerator/denominators?
-    /// TODO(dimo): Does Solidity allow literal operations that result in infinite fractional numbers? E.g. 2/3 ?
-    public readonly literal?: Decimal;
+    public readonly literal?: bigint;
 
-    constructor(literal?: bigint | Decimal, src?: Range) {
+    constructor(literal?: bigint, src?: Range) {
         super(src);
 
-        this.literal =
-            literal === undefined || literal instanceof Decimal
-                ? literal
-                : new Decimal(literal.toString());
-    }
-
-    negated(): IntLiteralType {
-        return new IntLiteralType(this.literal ? this.literal.negated() : this.literal);
+        this.literal = literal;
     }
 
     pp(): string {
-        if (this.literal && !this.literal.isInt()) {
-            const fraction = this.literal.toFraction();
-            return `rational_const ${fraction[0]} / ${fraction[1]}`;
-        }
-
-        return `int_const${this.literal !== undefined ? ` ${this.literal.toFixed()}` : ""}`;
+        return `int_const${this.literal !== undefined ? ` ${this.literal.toString()}` : ""}`;
     }
 
     /**
