+++ src/types/ast/numeric_literal.ts
@@ -1,0 +1,4 @@
+import { TypeNode } from "./type";
+
+/// Common base class for IntLiteralType and RationalLiteralType
+export abstract class NumericLiteralType extends TypeNode {}
