+++ src/types/eval_const.ts
@@ -1,0 +1,293 @@
+import Decimal from "decimal.js";
+import {
+    Expression,
+    Literal,
+    UnaryOperation,
+    BinaryOperation,
+    Conditional,
+    LiteralKind
+} from "../ast";
+import { pp } from "../misc";
+import { binaryOperatorGroups } from "./utils";
+
+export function isConstant(expr: Expression): boolean {
+    if (expr instanceof Literal) {
+        return true;
+    }
+
+    if (expr instanceof UnaryOperation && isConstant(expr.vSubExpression)) {
+        return true;
+    }
+
+    if (
+        expr instanceof BinaryOperation &&
+        isConstant(expr.vLeftExpression) &&
+        isConstant(expr.vRightExpression)
+    ) {
+        return true;
+    }
+
+    /// TODO: We can be more precise here. Conditionals are also constant if
+    /// 1) vCondition is constant, and only the selected branch is constant
+    /// 2) vCondition is not constant, but both branches are constant and equal (not likely in practice)
+    if (
+        expr instanceof Conditional &&
+        isConstant(expr.vCondition) &&
+        isConstant(expr.vTrueExpression) &&
+        isConstant(expr.vFalseExpression)
+    ) {
+        return true;
+    }
+
+    /// TODO: After eval add case for constant indexing
+    return false;
+}
+
+export type Value = Decimal | boolean | string | bigint;
+
+export class EvalError extends Error {
+    expr: Expression;
+
+    constructor(e: Expression, msg: string) {
+        super(msg);
+        this.expr = e;
+    }
+}
+
+export class NonConstantExpressionError extends EvalError {
+    constructor(e: Expression) {
+        super(e, `Found non-constant expression ${pp(e)} during constant evaluation`);
+    }
+}
+
+function evalLiteral(expr: Literal): Value {
+    if (expr.kind === LiteralKind.Bool) {
+        return expr.value === "true";
+    }
+
+    if (expr.kind === LiteralKind.String || expr.kind === LiteralKind.UnicodeString) {
+        return expr.value;
+    }
+
+    if (expr.kind === LiteralKind.HexString) {
+        return expr.hexValue;
+    }
+
+    const dec = new Decimal(expr.value);
+
+    if (dec.isInteger()) {
+        return BigInt(dec.toFixed());
+    }
+
+    return dec;
+}
+
+function evalUnary(expr: UnaryOperation): Value {
+    const subVal = evalConstantExpr(expr.vSubExpression);
+
+    if (expr.operator === "!") {
+        if (!(typeof subVal === "boolean")) {
+            throw new EvalError(
+                expr.vSubExpression,
+                `Expected a boolean in ${pp(expr.vSubExpression)} not ${subVal}`
+            );
+        }
+
+        return !subVal;
+    }
+
+    if (expr.operator === "~") {
+        if (typeof subVal === "bigint") {
+            return ~subVal;
+        }
+
+        throw new EvalError(
+            expr.vSubExpression,
+            `Expected a number in ${pp(expr.vSubExpression)} not ${subVal}`
+        );
+    }
+
+    if (expr.operator === "-") {
+        if (subVal instanceof Decimal) {
+            return subVal.negated();
+        }
+
+        if (typeof subVal === "bigint") {
+            return -subVal;
+        }
+
+        throw new EvalError(
+            expr.vSubExpression,
+            `Expected a number in ${pp(expr.vSubExpression)} not ${subVal}`
+        );
+    }
+
+    throw new Error(`NYI unary operator ${expr.operator}`);
+}
+
+function evalBinaryLogic(expr: BinaryOperation, lVal: Value, rVal: Value): Value {
+    const op = expr.operator;
+
+    if (!(typeof lVal === "boolean" && typeof lVal === "boolean")) {
+        throw new EvalError(expr, `${op} expects booleans not ${lVal} and ${rVal}`);
+    }
+
+    if (op === "&&") {
+        return lVal && rVal;
+    }
+
+    if (op === "||") {
+        return lVal || rVal;
+    }
+
+    throw new Error(`Unknown logic op ${op}`);
+}
+
+function promoteToDec(d: Value): Decimal {
+    if (!(typeof d === "bigint" || d instanceof Decimal)) {
+        throw new Error(`Expected number not ${d}`);
+    }
+
+    return d instanceof Decimal ? d : new Decimal(d.toString());
+}
+
+function demoteFromDec(d: Decimal): Decimal | bigint {
+    return d.isInt() ? BigInt(d.toFixed()) : d;
+}
+
+function evalBinaryEquality(expr: BinaryOperation, lVal: Value, rVal: Value): Value {
+    let equal: boolean;
+
+    if (typeof lVal === "string" || typeof rVal === "string") {
+        throw new EvalError(expr, `Comparison not allowed for strings ${lVal} and ${rVal}`);
+    }
+
+    if (lVal instanceof Decimal && rVal instanceof Decimal) {
+        equal = lVal.equals(rVal);
+    } else {
+        equal = lVal === rVal;
+    }
+
+    if (expr.operator === "==") {
+        return equal;
+    }
+
+    if (expr.operator === "!=") {
+        return !equal;
+    }
+
+    throw new Error(`Unknown equality op ${expr.operator}`);
+}
+
+function evalBinaryComparison(expr: BinaryOperation, lVal: Value, rVal: Value): Value {
+    const op = expr.operator;
+
+    const lDec = promoteToDec(lVal);
+    const rDec = promoteToDec(rVal);
+
+    if (op === "<") {
+        return lDec.lessThan(rDec);
+    }
+
+    if (op === "<=") {
+        return lDec.lessThanOrEqualTo(rDec);
+    }
+
+    if (op === ">") {
+        return lDec.greaterThan(rDec);
+    }
+
+    if (op === ">=") {
+        return lDec.greaterThanOrEqualTo(rDec);
+    }
+
+    throw new Error(`Unknown comparison op ${expr.operator}`);
+}
+
+function evalBinaryArithmetic(expr: BinaryOperation, lVal: Value, rVal: Value): Value {
+    const op = expr.operator;
+
+    const lDec = promoteToDec(lVal);
+    const rDec = promoteToDec(rVal);
+    let res: Decimal;
+
+    if (op === "+") {
+        res = lDec.plus(rDec);
+    } else if (op === "-") {
+        res = lDec.minus(rDec);
+    } else if (op === "*") {
+        res = lDec.times(rDec);
+    } else if (op === "/") {
+        res = lDec.div(rDec);
+    } else if (op === "%") {
+        res = lDec.modulo(rDec);
+    } else if (op === "**") {
+        res = lDec.pow(rDec);
+    } else {
+        throw new Error(`Unknown arithmetic op ${expr.operator}`);
+    }
+
+    return demoteFromDec(res);
+}
+
+function evalBinaryBitwise(expr: BinaryOperation, lVal: Value, rVal: Value): Value {
+    throw new Error(`NYI Bitwise op ${expr.operator} on ${lVal} ${rVal}`);
+}
+
+/// Given a constant `BinaryOperation` expression `expr` evaluate it.
+function evalBinary(expr: BinaryOperation): Value {
+    const lVal = evalConstantExpr(expr.vLeftExpression);
+    const rVal = evalConstantExpr(expr.vRightExpression);
+
+    if (binaryOperatorGroups.Logical.includes(expr.operator)) {
+        return evalBinaryLogic(expr, lVal, rVal);
+    }
+
+    if (binaryOperatorGroups.Equality.includes(expr.operator)) {
+        return evalBinaryEquality(expr, lVal, rVal);
+    }
+
+    if (binaryOperatorGroups.Comparison.includes(expr.operator)) {
+        return evalBinaryComparison(expr, lVal, rVal);
+    }
+
+    if (binaryOperatorGroups.Arithmetic.includes(expr.operator)) {
+        return evalBinaryArithmetic(expr, lVal, rVal);
+    }
+
+    if (binaryOperatorGroups.Bitwise.includes(expr.operator)) {
+        return evalBinaryBitwise(expr, lVal, rVal);
+    }
+
+    throw new Error(`Unknown binary op ${expr.operator}`);
+}
+
+/**
+ * Given a constant expression `expr` evaluate it to a concrete `Value`. If `expr`
+ * is not constant throw `NonConstantExpressionError`.
+ *
+ * TODO: The order of some operations changed in some version. So perhaps to be fully
+ * precise here we will need a compiler version too?
+ */
+export function evalConstantExpr(expr: Expression): Value {
+    if (expr instanceof Literal) {
+        return evalLiteral(expr);
+    }
+
+    if (expr instanceof UnaryOperation && isConstant(expr.vSubExpression)) {
+        return evalUnary(expr);
+    }
+
+    if (
+        expr instanceof BinaryOperation &&
+        isConstant(expr.vLeftExpression) &&
+        isConstant(expr.vRightExpression)
+    ) {
+        return evalBinary(expr);
+    }
+
+    /// Note that from the point of view of the type system constant conditionals and
+    /// indexing in constant array literals are not considered constant expressions.
+    /// So for now we don't support them, but we may change that in the future.
+    throw new Error(`${pp(expr)} is not a constant expression.`);
+}
