+++ src/types/index.ts
@@ -5,3 +5,4 @@ export * from "./utils";
 export * from "./infer";
 export * from "./polymorphic";
 export * from "./reserved";
+export * from "./eval_const";
